third_party/WebKit/LayoutTests/external/wpt/html/dom/documents/dom-tree-accessors/document.getElementsByName/document.getElementsByName-namespace-xhtml-expected.txt

This is a testharness.js-based test.
FAIL getElementsByName and foreign namespacesassert_equals: expected 1 but got 2
Harness: the test ran to completion.
